InvoicePlaceholder/requirements.txt
pyenchant==3.1.1
openpyxl==3.0.4
fastapi==0.60.1
uvicorn==0.11.8
XlsxWriter==1.3.2